+++ options_test.go
@@ -1,0 +1,21 @@
+package centrifuge
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestSkipHistory(t *testing.T) {
+	opt := SkipHistory()
+	opts := &PublishOptions{}
+	opt(opts)
+	assert.Equal(t, true, opts.SkipHistory)
+}
+
+func TestWithResubscribe(t *testing.T) {
+	opt := WithResubscribe()
+	opts := &UnsubscribeOptions{}
+	opt(opts)
+	assert.Equal(t, true, opts.Resubscribe)
+}
